Clean up code and refactor sanity check
